+++ src/scripts/klamak.ts
@@ -5,7 +5,7 @@ namespace Scripts {
 
         /**
          * Scripts.Klamak.next
-         * stability beta
+         * stability implementation
          *
          * x
          */
@@ -24,6 +24,74 @@ namespace Scripts {
 
             const nextList = Orion.GetGlobal('klamak');
             Scripts.Utils.playerPrint(nextList);
+        }
+
+        static useKlamak(lvl:number, useAim = false, priorityList:string[] = []) {
+            const level = gameObject.klamak['lvl' + lvl];
+
+            let findSerial = '';
+            for (const klamakName of priorityList) {
+                for (const itemName in level) {
+                    if (klamakName === itemName) {
+                        const klamak = level[itemName];
+                        const klamakSerials = Orion.FindType(klamak.graphic, klamak.color);
+                        if (klamakSerials) {
+                            findSerial = klamakSerials[0];
+                            break;
+                        }
+                    }
+                }
+                if (findSerial !== '') {
+                    break;
+                }
+            }
+
+            // no priority list or no klamak found by priority list
+            if (findSerial === '') {
+                for (const itemName in level) {
+                    if (priorityList.indexOf(itemName) > -1) {
+                        // exclude what was iterated before
+                        continue;
+                    }
+                    const klamak = level[itemName];
+                    const klamakSerials = Orion.FindType(klamak.graphic, klamak.color);
+                    if (klamakSerials.length) {
+                        findSerial = klamakSerials[0];
+                        break;
+                    }
+                }
+            }
+
+            if (findSerial === '') {
+                Scripts.Utils.playerPrint(`[ nemas pety ]`, ColorEnum.red);
+                return;
+            }
+
+            if (useAim) {
+                const selection = Orion.WaitForAddObject('klamakTarget');
+                if (selection === 0) {
+                    return;
+                }
+                const target:any = {}
+
+                if (selection === 1) {
+                    const targetGameObject = Orion.FindObject('klamakTarget');
+                    target.x = targetGameObject.X();
+                    target.y = targetGameObject.Y();
+                    target.z = targetGameObject.Z();
+                }
+                else {
+                    target.x = SelectedTile.X();
+                    target.y = SelectedTile.Y();
+                    target.z = SelectedTile.Z();
+                }
+                Orion.MoveItem(findSerial, 1, "ground", target.x, target.y, target.z);
+                Orion.Wait(responseDelay);
+
+            }
+            Orion.WarMode(true);
+            Orion.Wait(100);
+            Orion.UseObject(findSerial);
         }
     }
 }